Merging r259682: ------------------------------------------------------------------------ r259682 | dexonsmith | 2016-02-03 11:30:20 -0800 (Wed, 03 Feb 2016) | 7 lines
re.results.form: Format out-of-range subexpression references as null

Rather than crashing in match_results::format() when a reference to a
marked subexpression is out of range, format the subexpression as empty
(i.e., replace it with an empty string).  Note that
match_results::operator[]() has a range-check and returns a null match
in this case, so this just re-uses that logic.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/libcxx/branches/release_38@260983 91177308-0d34-0410-b5e6-96231b3b80d8