[libc++][CMake] Do not define `cxx_shared_EXPORTS` when building the shared library
CMake will define -Dcxx_shared_EXPORTS when building the shared library
by default. In theory, this is used to signal to the library that we're
building a shared library and that dllimport/dllexport should be used.
However, we already have our own way of doing that, so I'm removing this
define to avoid meaningless command line arguments in the build.

git-svn-id: https://llvm.org/svn/llvm-project/libcxx/trunk@356167 91177308-0d34-0410-b5e6-96231b3b80d8