[libc++] Run back-deployment CI against previously-released libc++abi dylibs
We used to do it against the current system's libc++abi, which is not as
good as doing it with the libc++abi that matches the libc++ we're running
against.

Note that I made sure we were indeed picking up the provided libc++abi
by replacing it by something that doesn't work and watching it burn.

git-svn-id: https://llvm.org/svn/llvm-project/libcxx/trunk@358294 91177308-0d34-0410-b5e6-96231b3b80d8