Optimize std::midpoint
Same idea as the current algorithm, that is, add (half of the difference between a and b) to a.

But we use a different technique for computing the difference: we compute b - a into a pair of integers that are named "__sign_bit" and "__diff".  We have to use two because, for example, subtracting two 32-bit integers produces a 33-bit result.

Computing half of that is a simple matter of shifting __diff right by 1, and adding __sign_bit shifted left by 31.

The only tricky part is that if the difference is odd and negative, then shifting it by one isn't the same as dividing it by two - shifting a negative one produces a negative one, for example.  So there's one more adjustment: if the sign bit and the low bit of diff are one, we add one.

Here's a demonstration of the codegen difference, as well as some testing: https://godbolt.org/z/7ar3K9